Added updates from Tuomo Soini.

+++ Makefile
@@ -17,7 +17,6 @@ all:
 	@echo "Nothing to be build."
 
 install: dconf.1
-#	-[ ! -f $(DESTDIR)$(sysconfdir)/dconf.conf ] && install -D -m0644 dconf.conf $(DESTDIR)$(sysconfdir)/dconf.conf
 	-@[ ! -f $(DESTDIR)$(sysconfdir)/dconf-custom.conf ] && install -D -m0644 config/dconf-custom.conf $(DESTDIR)$(sysconfdir)/dconf-custom.conf
 
 	install -Dp -m0755 dconf $(DESTDIR)$(bindir)/dconf
@@ -26,14 +25,14 @@ install: dconf.1
 	install -dp -m0755 $(DESTDIR)$(logdir)
 	@echo "Also do: make install-<dist>   (with dist: debian|redhat|suse)"
 
-install-redhat:
-	install -Dp -m0644 config/dconf-redhat.conf $(DESTDIR)$(sysconfdir)/dconf.conf
+install-redhat: install
+	-[ ! -f $(DESTDIR)$(sysconfdir)/dconf.conf ] && install -Dp -m0644 config/dconf-redhat.conf $(DESTDIR)$(sysconfdir)/dconf.conf
 
-install-debian:
-	install -Dp -m0644 config/dconf-debian.conf $(DESTDIR)$(sysconfdir)/dconf.conf
+install-debian: install
+	-[ ! -f $(DESTDIR)$(sysconfdir)/dconf.conf ] && install -Dp -m0644 config/dconf-debian.conf $(DESTDIR)$(sysconfdir)/dconf.conf
 
 install-suse:
-	install -Dp -m0644 config/dconf-suse.conf $(DESTDIR)$(sysconfdir)/dconf.conf
+	-[ ! -f $(DESTDIR)$(sysconfdir)/dconf.conf ] && install -Dp -m0644 config/dconf-suse.conf $(DESTDIR)$(sysconfdir)/dconf.conf
 
 clean:
 	rm -f dconf.1 dconf.1.html dconf.1.xml
